+++ examples/09_Widgets/src/WidgetsExample.cpp
@@ -10,7 +10,7 @@ class Action1 : public iAction
 
 public:
     Action1()
-        : iAction("example:one")
+        : iAction("exampleAction:one")
     {
         setDescription("Action One");
         setIcon("igor_icon_camera");
@@ -41,7 +41,7 @@ class Action2 : public iAction
 
 public:
     Action2()
-        : iAction("example:two")
+        : iAction("exampleAction:two")
     {
         setDescription("Action Two");
     }
@@ -71,7 +71,7 @@ class Action3 : public iAction
 
 public:
     Action3()
-        : iAction("example:three")
+        : iAction("exampleAction:three")
     {
         setIcon("igor_icon_delete");
     }
@@ -101,7 +101,7 @@ class Action4 : public iAction
 
 public:
     Action4()
-        : iAction("example:four")
+        : iAction("exampleAction:four")
     {
         setIcon("igor_icon_delete");
     }
@@ -155,10 +155,10 @@ iWidgetMenuBarPtr WidgetsExample::createMenu()
     iWidgetMenuBarPtr menuBar = new iWidgetMenuBar();
 
     // get some actions to add to the menu
-    iActionPtr action1 = iActionManager::getInstance().getAction("example:one");
-    iActionPtr action2 = iActionManager::getInstance().getAction("example:two");
-    iActionPtr action3 = iActionManager::getInstance().getAction("example:three");
-    iActionPtr action4 = iActionManager::getInstance().getAction("example:four");
+    iActionPtr action1 = iActionManager::getInstance().getAction("exampleAction:one");
+    iActionPtr action2 = iActionManager::getInstance().getAction("exampleAction:two");
+    iActionPtr action3 = iActionManager::getInstance().getAction("exampleAction:three");
+    iActionPtr action4 = iActionManager::getInstance().getAction("exampleAction:four");
 
     // build up menu and submenus
     iWidgetMenuPtr menu1 = new iWidgetMenu();
@@ -210,53 +210,54 @@ void WidgetsExample::onInit()
     iLayerWidgets::onInit();
 
     _dialog = new iDialog();
+    _dialog->setBorder(5);
     _dialog->setPos(40, getWindow()->getClientHeight() - 420);
 
-    iWidgetGrid *gridMain = new iWidgetGrid(_dialog);
-    gridMain->appendRows(3);
-    gridMain->setHorizontalAlignment(iHorizontalAlignment::Stretch);
-    gridMain->setVerticalAlignment(iVerticalAlignment::Stretch);
-    gridMain->setBorder(10);
-    gridMain->setStretchRow(2);
+    iWidgetBoxLayoutPtr mainLayout = new iWidgetBoxLayout(iWidgetBoxLayoutType::Vertical, _dialog);
 
-    // create a menu and add to grid
+    // create a menu and add to layout
     iWidgetMenuBarPtr menuBar = createMenu();
-    gridMain->addWidget(menuBar, 0, 0);
+    mainLayout->addWidget(menuBar);
+
+    // add some buttons
+    iWidgetBoxLayoutPtr buttonLayout = new iWidgetBoxLayout(iWidgetBoxLayoutType::Horizontal);
+    mainLayout->addWidget(buttonLayout);
+
+    iWidgetButton *exitButton = new iWidgetButton();
+    exitButton->setAction(iActionManager::getInstance().getAction("igor:exit"));
+    exitButton->setText(""); // clear text
+    exitButton->setSize(24, 24);
+    buttonLayout->addWidget(exitButton);
+
+    iWidgetButton *button2 = new iWidgetButton();
+    button2->setAction(iActionManager::getInstance().getAction("exampleAction:one"));
+    button2->setText(""); // clear text
+    button2->setSize(24, 24);
+    buttonLayout->addWidget(button2);
+
+    iWidgetButton *button3 = new iWidgetButton();
+    button3->setAction(iActionManager::getInstance().getAction("exampleAction:two"));
+    button3->setSize(24, 24);
+    buttonLayout->addWidget(button3);
 
     // adding a group box
     iWidgetGroupBox *groupBox1 = new iWidgetGroupBox();
     groupBox1->setText("Hello World. This is a group box!");
     groupBox1->setHorizontalAlignment(iHorizontalAlignment::Stretch);
-    gridMain->addWidget(groupBox1, 0, 1);
+    mainLayout->addWidget(groupBox1);
 
-    // add some buttons
-    iWidgetGrid *gridButtons = new iWidgetGrid();
-    gridButtons->appendColumns(3);
-    gridButtons->setStretchColumn(3);
-    gridButtons->setHorizontalAlignment(iHorizontalAlignment::Stretch);
-    gridMain->addWidget(gridButtons, 0, 1);
+    iWidgetBoxLayoutPtr vBoxLayout2 = new iWidgetBoxLayout(iWidgetBoxLayoutType::Vertical);
+    groupBox1->addWidget(vBoxLayout2);
 
-    iWidgetButton *exitButton = new iWidgetButton();
-    exitButton->setAction(iActionManager::getInstance().getAction("igor:exit"));
-    exitButton->setText("");
-    exitButton->setVerticalTextAlignment(iVerticalAlignment::Bottom);
-    exitButton->setVerticalAlignment(iVerticalAlignment::Center);
-    exitButton->setHorizontalAlignment(iHorizontalAlignment::Center);
-    exitButton->setSize(24, 24);
-    gridButtons->addWidget(exitButton, 0, 0);
-
-    // controls on the left
-    iWidgetGrid *gridControls = new iWidgetGrid();
-    gridControls->appendColumns(1);
-    gridControls->appendRows(3);
-    gridControls->setHorizontalAlignment(iHorizontalAlignment::Stretch);
-    gridControls->setVerticalAlignment(iVerticalAlignment::Top);
-    gridMain->addWidget(gridControls, 0, 2);
+    iWidgetBoxLayoutPtr profileLayout = new iWidgetBoxLayout(iWidgetBoxLayoutType::Horizontal);
+    vBoxLayout2->addWidget(profileLayout);
 
     iWidgetLabel *labelProfiler = new iWidgetLabel();
     labelProfiler->setText("Profiler:");
     labelProfiler->setHorizontalAlignment(iHorizontalAlignment::Right);
-    gridControls->addWidget(labelProfiler, 0, 0);
+    labelProfiler->setVerticalAlignment(iVerticalAlignment::Top);
+    labelProfiler->setWidth(80);
+    profileLayout->addWidget(labelProfiler);
 
     iWidgetSelectBoxPtr selectBox = new iWidgetSelectBox();
     selectBox->addItem("None");
@@ -267,25 +268,38 @@ void WidgetsExample::onInit()
     selectBox->setSelection((int)_layerProfiler->getVerbosity());
     selectBox->setHorizontalAlignment(iHorizontalAlignment::Right);
     selectBox->registerOnChangeEvent(iChangeDelegate(this, &WidgetsExample::onProfilerVerbosityChange));
-    gridControls->addWidget(selectBox, 1, 0);
+    profileLayout->addWidget(selectBox);
+
+    iWidgetBoxLayoutPtr optionsLayout = new iWidgetBoxLayout(iWidgetBoxLayoutType::Horizontal);
+    vBoxLayout2->addWidget(optionsLayout);
+
+    iWidgetLabel *labelOptions = new iWidgetLabel();
+    labelOptions->setText("Options:");
+    labelOptions->setHorizontalAlignment(iHorizontalAlignment::Right);
+    labelOptions->setVerticalAlignment(iVerticalAlignment::Top);
+    labelOptions->setWidth(80);
+    optionsLayout->addWidget(labelOptions);    
+
+    iWidgetBoxLayoutPtr radioLayout = new iWidgetBoxLayout(iWidgetBoxLayoutType::Vertical);
+    optionsLayout->addWidget(radioLayout);
 
     iWidgetCheckBox *radioWireframe = new iWidgetCheckBox();
     radioWireframe->setText("Wireframe");
     radioWireframe->setHorizontalAlignment(iHorizontalAlignment::Left);
     radioWireframe->registerOnChangeEvent(iChangeDelegate(this, &WidgetsExample::onWireframeChange));
-    gridControls->addWidget(radioWireframe, 1, 1);
+    radioLayout->addWidget(radioWireframe);
 
     iWidgetCheckBox *radioOctree = new iWidgetCheckBox();
     radioOctree->setText("Octree");
     radioOctree->setHorizontalAlignment(iHorizontalAlignment::Left);
     radioOctree->registerOnChangeEvent(iChangeDelegate(this, &WidgetsExample::onOctreeChange));
-    gridControls->addWidget(radioOctree, 1, 2);
+    radioLayout->addWidget(radioOctree);
 
     iWidgetCheckBox *radioBoundings = new iWidgetCheckBox();
     radioBoundings->setText("Boundings");
     radioBoundings->setHorizontalAlignment(iHorizontalAlignment::Left);
     radioBoundings->registerOnChangeEvent(iChangeDelegate(this, &WidgetsExample::onBoundsChange));
-    gridControls->addWidget(radioBoundings, 1, 3);
+    radioLayout->addWidget(radioBoundings);
 
     _dialog->open(iDialogCloseDelegate(this, &WidgetsExample::onCloseDialog));