+++ src/igor/igor/ui/dialogs/iDialogColorChooser.cpp
@@ -8,7 +8,7 @@
 #include <igor/ui/iWidgetManager.h>
 #include <igor/ui/widgets/iWidgetLabel.h>
 #include <igor/ui/widgets/iWidgetButton.h>
-#include <igor/ui/widgets/iWidgetGrid.h>
+#include <igor/ui/widgets/iWidgetGridLayout.h>
 #include <igor/ui/widgets/iWidgetSpacer.h>
 #include <igor/ui/widgets/iWidgetCheckBox.h>
 #include <igor/ui/iWidgetManager.h>
@@ -41,7 +41,7 @@ namespace igor
         setWidth(20);
         setHeight(20);
 
-        iWidgetGridPtr grid = new iWidgetGrid(this);
+        iWidgetGridLayoutPtr grid = new iWidgetGridLayout(this);
         grid->appendRows(2);
         grid->setHorizontalAlignment(iHorizontalAlignment::Center);
         grid->setVerticalAlignment(iVerticalAlignment::Center);
@@ -58,7 +58,7 @@ namespace igor
         _userControlColorChooser->setHeadlineVisible(false);
         _userControlColorChooser->setColor(color);
 
-        iWidgetGrid *buttonGrid = new iWidgetGrid();
+        iWidgetGridLayout *buttonGrid = new iWidgetGridLayout();
         buttonGrid->appendColumns(2);
         buttonGrid->setHorizontalAlignment(iHorizontalAlignment::Right);