+++ src/tools/mica/src/usercontrols/UserControlEmitter.cpp
@@ -59,7 +59,7 @@ uint32 UserControlEmitter::getNode()
 
 void UserControlEmitter::initGUI()
 {
-    _grid = new iWidgetGrid();
+    _grid = new iWidgetGridLayout();
     _grid->appendColumns(1);
     _grid->appendRows(1);
     _grid->setHorizontalAlignment(iHorizontalAlignment::Stretch);